Light Linger
- created "InteractWithLight" script that determines how things like
walls, etc., interact with light. When set to "Absorb" a "LightLinger"
object is created where the bullet hits the surface
- "LightLinger" objects fade over time, can be used to light up an area
for a few moments (5s right now)
- "DebugLight" is just a directional light we can place to help us see
things while designing a level
- I was getting some errors in the "DetectHitOnObject" class where the
bullet script was null. I put in a line of code to return the function
when that happens rather than throw an error.