Puzzle 2 implementation complete.
The player enters a button, in this case Nathan's prism, and buttons related targets/sensors are made visible. They have a light. When they are shot with the correct light frequency (Which in this case is just the one that the player gets when they step into the prism) the light is turned off.

After all the lights in all the targets/sensors are turned off. The player has completed the puzzle. You can see this success statement in DetectPuz2.cs